Create structs for internal annotation tuples (#404)
Refactors several tuples in the main annotation function into structs.
Creates a generic implementation of the vec_utils::find_next* functions, re-implements all of the existing functions using the generic version, and fixes the input and output types to be usize (and fixes all the calling locations).

+++ enclone_print/src/print_clonotypes.rs
@@ -229,7 +229,7 @@ pub fn print_clonotypes(
         let mut mults = Vec::<usize>::new();
         let mut j = 0;
         while j < od.len() {
-            let k = next_diff12_3(&od, j as i32) as usize;
+            let k = next_diff12_3(&od, j);
             let mut mult = 0_usize;
             for l in j..k {
                 let x: &CloneInfo = &info[od[l].2 as usize];
